requirements.txt
cryptography
psutil
pyjwt
requests
opencv-python
slack_sdk
